ai/pyproject.toml
[tool.poetry]
name = "ai_server"
version = "1.10.0"
description = "Vika AI"
authors = ["APITable Ltd. <dev@apitable.com>"]
license = "COMMERCIAL"
readme = "README.md"
packages = [{ include = "ai_server" }]

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.109.0"
python-dotenv = "^1.0.0"
bs4 = "^0.0.1"
chromadb = "^0.3.26"
requests = "^2.31.0"
inflection = "^0.5.1"
sse-starlette = "^1.8.2"
frozendict = "^2.4.0"
python-dateutil = "^2.8.2"
typing-extensions = "^4.9.0"
urllib3 = "^2.1.0"
loguru = "^0.7.0"
httpx = "^0.26.0"
html2text = "^2020.1.16"
pydantic = "1.10.13"
unstructured = "^0.12.0"
markdown = "^3.5.2"
jieba = "^0.42.1"
sentry-sdk = "^1.39.2"
pypdf = "^3.17.4"
pdf2image = "^1.17.0"
pdfminer-six = "^20221105"
unstructured-inference = "^0.7.21"
pytesseract = "^0.3.10"
docx2txt = "^0.8"
python-docx = "^1.1.0"
socksio = "^1.0.0"
jsonschema = "^4.21.0"
pyairtable = "^2.2.1"
apitable = "1.3.1"
aiofiles = "^23.2.1"
openai = "^1.8.0"
qianfan = "0.1.1"
aiohttp = "^3.9.1"
langchain = "^0.1.1"
tiktoken = "^0.5.2"
langchain-openai = "^0.0.2"
pymupdf = "^1.23.15"

[tool.poetry.group.dev.dependencies]
setuptools = "^68.0.0"
ipykernel = "^6.23.2"
pytest = "^7.4.0"
fake-useragent = "^1.1.3"
pytest-asyncio = "^0.21.0"
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.277"
pytest-loguru = "^0.2.0"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
test_path = ["tests"]

[tool.black]
line-length=90

[tool.isort]
profile = "black"
line_length=90
multi_line_output=3

[tool.ruff]
line-length=90
ignore = [
    # E501: line-too-long, ref: https://beta.ruff.rs/docs/rules/line-too-long/
    "E501"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
